core/wsgi.py
"""
WSGI config for missinglinks project.

It exposes the WSGI callable as a module-level variable named ``application``.

For more information on this file, see
https://docs.djangoproject.com/en/1.8/howto/deployment/wsgi/
"""
import environ
path = environ.Path(__file__) - 2
environ.Env().read_env(path('.env'))

from django.core.wsgi import get_wsgi_application
from dj_static import Cling, MediaCling

application = Cling(MediaCling(get_wsgi_application()))
