core/requirements.txt
Django>=1.8,<1.9
wagtail==1.1
django-environ==0.4.0
psycopg2==2.6.1
dj-static==0.0.6
gunicorn==19.3.0
